+++ src/main/java/mpern/sap/commerce/ccv2/tasks/ValidateManifest.java
@@ -43,6 +43,7 @@ public class ValidateManifest extends DefaultTask {
         ManifestExtensionsResolver resolver = new ManifestExtensionsResolver(projectDir);
         validators.add(new CloudHotfolderValidator(projectDir, resolver));
         validators.add(new MediaConversionValidator(resolver));
+        validators.add(new WebrootValidator(projectDir));
         boolean deepInspection = false;
         if (Files.exists(projectDir.resolve("hybris/bin/platform"))) {
             deepInspection = true;