+++ CHANGELOG.md
@@ -10,8 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Task `validateManifest`
-  
+
   - Check setup of cloud media conversion
+  - Check if `<extension>.webroot` properties are present
 
 ## [3.2.0] 2021-01-15